refactor: modal 생성 함수 분리
